refactor test_runner.rs as a directory
